Cleaned Up code, Fixed some Errors; Basically Final Version
- Cleaned up code commenting, rearrange variables, names, types etc.
-  Fixed Health generation code as previously genrated negative Health amounts.

+++ Source/AdvGamesProgramming/HealthComponent.cpp
@@ -61,7 +61,7 @@ float UHealthComponent::HealthPercentageRemaining()
 	return CurrentHealth / MaxHealth; 
 }
 
-void UHealthComponent::OnTouchHealthBoost(int32 HealthAmount)
+void UHealthComponent::OnTouchHealthBoost(float HealthAmount)
 {
 	CurrentHealth += HealthAmount; 
 }
