+++ Source/AdvGamesProgramming/HealthComponent.h
@@ -38,5 +38,5 @@ public:
 	UFUNCTION(BlueprintCallable)
 	void OnDeath();
 
-	void OnTouchHealthBoost(int32 HealthAmount);
+	void OnTouchHealthBoost(float HealthAmount);
 };
